Move SignUpForm and icon files
