+++ src/app/auth/common/components/icons/email.tsx
