+++ src/app/auth/common/components/icons/key.tsx
